+++ src/test/java/org/ttzero/excel/entity/csv/ListObjectSheetTest.java
@@ -23,7 +23,6 @@ import org.ttzero.excel.entity.Column;
 import org.ttzero.excel.entity.CustomizeDataSourceSheet;
 import org.ttzero.excel.entity.ExcelWriteException;
 import org.ttzero.excel.entity.ListSheet;
-import org.ttzero.excel.entity.Sheet;
 import org.ttzero.excel.entity.TooManyColumnsException;
 import org.ttzero.excel.entity.Workbook;
 import org.ttzero.excel.entity.WorkbookTest;
@@ -101,7 +100,7 @@ public class ListObjectSheetTest extends WorkbookTest{
             .addSheet(Student.randomTestData()
                 , new Column("学号", "id")
                 , new Column("姓名", "name")
-                , new Column("成绩", "score", n -> n < 60 ? "不及格" : n)
+                , new Column("成绩", "score", n -> (int) n < 60 ? "不及格" : n)
             )
             .saveAsCSV()
             .writeTo(getOutputTestPath());
@@ -132,9 +131,9 @@ public class ListObjectSheetTest extends WorkbookTest{
             .addSheet(Student.randomTestData()
                 , new Column("学号", "id")
                 , new Column("姓名", "name")
-                , new Column("成绩", "score", n -> n < 60 ? "不及格" : n)
+                , new Column("成绩", "score", n -> (int) n < 60 ? "不及格" : n)
                     .setStyleProcessor((o, style, sst) -> {
-                        if ((int)o < 60) {
+                        if ((int) o < 60) {
                             style = Styles.clearFill(style)
                                 | sst.addFill(new Fill(PatternType.solid, new Color(246, 209, 139)));
                         }
